Refactor file loading code
Fixes bugs caused by the introduction of youtube playlist.
When a youtube playlist is loaded, opening a video from the playlist
should not update the lastPlayed setting, else on next run it would
load a single video instead of the playlist. Loading the correct
playlist video on startup is done through the lastPlaylistIndex setting